requirements.txt
selenium
pyyaml
pytz
pathlib